wip: prep for l11 - doctrine breaking changes
