+++ src/Support/ResponseExtractor/ModelInfoProviders/NativeProvider.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Dedoc\Scramble\Support\ResponseExtractor\ModelInfoProviders;
+
+class NativeProvider
+{
+
+}
